Moves HttpRequest struct into its own header, renames Response to HttpResponse
