+++ includes/Client.hpp
@@ -2,8 +2,9 @@
 
 #include <memory>
 #include <ctime>
+#include "HttpRequest.hpp"
 #include "RequestHandler.hpp"
-#include "Response.hpp"
+#include "HttpResponse.hpp"
 #include "ResponseHandler.hpp"
 
 class ResponseHandler;
@@ -13,9 +14,9 @@ struct Client {
 	std::string							requestString;
 	bool								requestReady;
 	std::unique_ptr<HttpRequest>		request;
-	std::unique_ptr<Response>			response;
+	std::unique_ptr<HttpResponse>			response;
 	std::unique_ptr<ResponseHandler>	responseHandler;
-	// std::unique_ptr<RequestHandler>		requestHandler;
+	std::unique_ptr<RequestHandler>		requestHandler;
 	int									fileSize;
 	int									fileReadFd;
 	int									fileTotalBytesRead;
